Add GenericModelSerializer
Which is basically the same as GenericRelatedField, but it allows you to serialize
a list of top-level objects with disparate types.
